+++ source/extensions/omni.flux.stage_manager.factory/omni/flux/stage_manager/factory/plugins/__init__.py
@@ -20,6 +20,7 @@ __all__ = [
     "StageManagerContextPlugin",
     "StageManagerFilterPlugin",
     "StageManagerInteractionPlugin",
+    "StageManagerListenerPlugin",
     "StageManagerTreePlugin",
     "StageManagerWidgetPlugin",
     "LengthUnit",
@@ -30,5 +31,6 @@ from .column_plugin import ColumnWidth, LengthUnit, StageManagerColumnPlugin
 from .context_plugin import StageManagerContextPlugin
 from .filter_plugin import StageManagerFilterPlugin
 from .interaction_plugin import StageManagerInteractionPlugin
+from .listener_plugin import StageManagerListenerPlugin
 from .tree_plugin import StageManagerTreePlugin
 from .widget_plugin import StageManagerWidgetPlugin