+++ CHANGELOG.md
@@ -28,6 +28,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - REMIX-3441: Implemented Lights interaction plugins
 - REMIX-3214: Checking layer type at project file import validation
 - REMIX-3215: Checking layer type at mod file import validation
+- REMIX-3477: Implemented USD listeners for the stage manager
+- REMIX-3478: Implemented selection syncing for the stage manager
 
 ### Changed