+++ source/extensions/omni.flux.stage_manager.plugin.tree.usd/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.3.1]
+### Fixed
+- Clear the items on refresh before re-adding new items
+
 ## [1.3.0]
 ### Added
 - Added `LightGroupsTreePlugin` plugin