+++ source/extensions/omni.flux.stage_manager.widget/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.0.1]
+### Fixed
+- Added `handle_exception` decorator for async functions
+
 ## [1.0.0]
 ### Added
 - Init commit.