+++ source/extensions/omni.flux.stage_manager.window/docs/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.0.2]
+### Added
+- Added USD listener extension as dependency for the standalone app
+
+### Changed
+- Ignore cameras prim paths in the stage manager schema
+
 ## [1.0.1]
 ### Changed
 - Updated test schema to include the Light interaction plugin